+++ CHANGELOG.md
@@ -2,5 +2,6 @@
 - 
 ### Fixed:
 - Fixed crash with blast furnace disabled slot
+- Fixed dehydration compat
 ### Changed:
 -